feat: interceptors integrated for uiSchema
