+++ generator/config.js
@@ -1,0 +1,5 @@
+const CONFIG = {
+  modules: ['components', 'utils', 'interceptors'],
+};
+
+module.exports = CONFIG;
