+++ src/fields/configure/component/get-component-props.js
@@ -15,9 +15,11 @@ export default ({
   data,
   objectData,
   dynamicKeyField,
+  interceptors,
 }) => {
   const widget = uiSchema['ui:widget'];
   const options = uiSchema['ui:options'] || uiSchema['ui:props'] || {};
+  const interceptorFunc = uiSchema['ui:interceptor'] || options.onBeforeChange;
   const { type } = schema;
   const rv = isCustomComponent
     ? {
@@ -26,7 +28,19 @@ export default ({
       ...options,
     }
     : {
-      onChange,
+      onChange: (event, value) => {
+        // Call Interceptor if it exists
+        if (typeof interceptors[interceptorFunc] === 'function') {
+          console.log('interceptors normal about to be called', interceptors);
+          const { formData, uiData } = interceptors[interceptorFunc]({
+            value,
+            options,
+          });
+
+          return onChange(event, formData, uiData);
+        } 
+        return onChange(event, value);
+      },
       onKeyDown,
       uiSchema,
       schema,
@@ -48,7 +62,11 @@ export default ({
     && isCustomComponent({ onChange }).props
     && isCustomComponent({ onChange }).props.onChange
   ) {
-    rv.onChange = isCustomComponent({ onChange }).props.onChange;
+    rv.onChange = (event, value) => {
+      // Call Interceptor if it exists
+      console.log('interceptors custom about to be called', interceptors);
+      return isCustomComponent({ onChange }).props.onChange(event, value);
+    };
   }
   else if (options.disabled) {
     if (typeof options.disabled === 'boolean') {
