+++ src/fields/interceptors/translate-currency/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@react-jsonschema-form-interceptors/translate-currency",
+  "version": "1.0.0",
+  "author": "Vipin Tanna"
+}
