+++ src/fields/interceptors/translate-ratings/index.js
@@ -1,0 +1,10 @@
+const translateRatings = ({ value, options }) => {
+  const formData = parseFloat(value);
+  const uiData = Number(value);
+  return {
+    formData,
+    uiData,
+  };
+};
+
+export default translateRatings;
