+++ src/fields/interceptors/translate-ratings/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@react-jsonschema-form-interceptors/translate-ratings",
+  "version": "1.0.0",
+  "author": "Vipin Tanna"
+}
